backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
Pillow
imagehash
chromadb
invisible-watermark
google-api-python-client
yt-dlp
apify-client  
 